+++ CHANGELOG.md
@@ -76,3 +76,13 @@ the following form: YYYY.0M.0D.
 [unreleased]: https://github.com/jrg94/portfolio-project/compare/v2024.08.07...HEAD
 [2024.08.07]: https://github.com/jrg94/portfolio-project/compare/v2024.01.07...v2024.08.07
 [2024.01.07]: https://github.com/jrg94/portfolio-project/releases/tag/v2024.01.07
+
+## 2025.04.2
+
+### Added
+
+- Designed abstract class for CurrencyConverter component
+
+### Updated
+
+- Implemented enhanced methods using kernel interface